+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.3.3] - 2020-02-18
+
+### Added
+- Support for `.onion` Tor URLs without SSL certificate (both http and https are valid for `TorUrl`).
+
+### Fixed
+- `__repr__` mixin.
+
 ## [0.3.2] - 2020-01-31
 
 ### Added
@@ -81,7 +89,8 @@ All notable changes to this project will be documented in this file.
 - Initial commit, based on `bech32` package.
 - `encode()` and `decode()` functions.
 
-[unreleased]: https://github.com/python-ln/lnurl/compare/0.3.2...HEAD
+[unreleased]: https://github.com/python-ln/lnurl/compare/0.3.3...HEAD
+[0.3.3]: https://github.com/python-ln/lnurl/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/python-ln/lnurl/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/python-ln/lnurl/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/python-ln/lnurl/compare/0.2.0...0.3.0